pyproject.toml
[tool.poetry]
name = "asnakedeck"
version = "0.0.1"
description = "AsyncIO StreamDeck controller for Linux"
authors = ["Ash Berlin-Taylor <ash_github@firemirror.com>"]
license = "MIT"
readme = "README.md"
exclude = ["examples/"]

[tool.poetry.dependencies]
python = "^3.10"
streamdeck = ">= 0.8.4"
attrs = "*"
pillow = "*"
asyncinotify = "^2.0.2"

[tool.poetry.dev-dependencies]
vulture = "^1.0"
bandit = "^1.6"
safety = "^1.8"
isort = "^5.3"
flake8-bugbear = "^19.8"
black = {version = "*", allow-prereleases = true}
mypy = "~0.812"
ipython = "^7.7"
pytest = "^6.0"
pytest-cov = "^2.7"
pep8-naming = "^0.8.2"
cruft = "^2.2"
pytest-deadfixtures = "^2.2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 160
skip-string-normalization = true

[tool.isort]
balanced_wrapping = true
default_section = "THIRDPARTY"
include_trailing_comma = true
known_first_party = [ "asnakedeck", "tests" ]
line_length = 160
multi_line_output = 3

[tool.cruft]
skip = [
  "asnakedeck/__init__.py",
  "tests",
  "pyproject.toml",
  "poetry.lock",
  "README.md",
  "CHANGELOG.md",
  ".flake8",
  ".git",
]

[tool.poetry.plugins."asnakedeck.key_handler"]
"clock" = "asnakedeck.handlers.clock:Clock"
"label" = "asnakedeck.handlers.label:Label"
"emoji" = "asnakedeck.handlers.emoji:Emoji"
"cycle" = "asnakedeck.handlers.cycle:Cycle"
